+++ module/SidebarTab.mjs
@@ -1,0 +1,6 @@
+// eslint-disable-next-line no-undef
+export default class HeroSystem6eSideBarTab extends SideBarTab {
+    constructor(...args) {
+        super(...args);
+    }
+}